# Improvement to performance of tensor.sum