@ndgrigorian 2023-07-26
Currently giving incorrect results for some cases with axes:
```
In [1]: import dpctl.tensor as dpt, numpy as np

In [2]: x = dpt.ones((2, 4098, 4098), dtype="i2")

In [3]: x_np = dpt.asnumpy(x)

In [4]: dpt.sum(x, axis=(0, 1))
Out[4]: usm_ndarray([8704, 8704, 8704, ...,   0,   0,   0])

In [5]: np.sum(x_np, axis=(0, 1))
Out[5]: array([8196, 8196, 8196, ..., 8196, 8196, 8196])
```