@ndgrigorian 2023-08-18
> The timing now:
> 
> ```
> (dev_dpctl) opavlyk@opavlyk-mobl:~/repos/dpctl$ ipython
> Python 3.9.12 (main, Jun  1 2022, 11:38:51)
> Type 'copyright', 'credits' or 'license' for more information
> IPython 8.14.0 -- An enhanced Interactive Python. Type '?' for help.
> 
> In [1]: import dpctl.tensor as dpt
> 
> In [2]: x = dpt.ones((4096, 4096), dtype="f4")
> 
> In [3]: y = dpt.sum(x, axis=0)
> 
> In [4]: %time y = dpt.sum(x, axis=0)
> CPU times: user 25.2 ms, sys: 6.65 ms, total: 31.9 ms
> Wall time: 33.8 ms
> 
> In [5]: %time y = dpt.sum(x, axis=0)
> CPU times: user 10.6 ms, sys: 20.1 ms, total: 30.7 ms
> Wall time: 30 ms
> 
> In [6]: %time y = dpt.sum(x, axis=0)
> CPU times: user 13.4 ms, sys: 19 ms, total: 32.4 ms
> Wall time: 31.1 ms
> 
> In [7]: %time y = dpt.sum(x, axis=0)
> CPU times: user 10.4 ms, sys: 20.1 ms, total: 30.4 ms
> Wall time: 29.5 ms
> 
> In [8]: !git describe --tag
> 0.14.6dev2-7-g9f544288f
> ```
> 
> It still has improved over 38 ms, but my claim of reducing to 4.7ms was erroneous (it was faster, due to it being incorrect).
> 
> @ndgrigorian I would prefer to merge this change, and defer further work to later PRs.

I agree. I'll approve this PR.